+++ src/ui/layout/Section.tsx
@@ -5,39 +5,51 @@ import Typography from '../atoms/Typography'
 import Button from '../atoms/Button'
 import Row from './Row'
 
+import styled from '@emotion/styled'
+
 interface Props {
   children: ReactNode | ReactElement | ReactElement[]
   title: string | ReactElement
   description?: string
   rightChild?: string | ReactElement
+  className?: string
 }
 
-function Section({ title, description, rightChild, children }: Props) {
+function Section({ title, description, rightChild, children, className }: Props) {
   return (
-    <section>
-      <Row align='space-between' styles={sectionTitleStyles}>
-        {typeof title === 'string' ?
-          <Typography type='title-3' size={20}>{title}</Typography>
-          : title}
-        <div>{rightChild}</div>
-      </Row>
-      <Typography type='description'>{description}</Typography>
+    <Container className={className}>
+      <Header>
+        <Row align='space-between' styles={sectionTitleStyles}>
+          {typeof title === 'string' ?
+            <Typography type='title-3' size={16}>{title}</Typography>
+            : title}
+          <div>{rightChild}</div>
+        </Row>
+        <Typography type='description'>{description}</Typography>
+      </Header>
 
       {children}
-    </section>
+    </Container>
   )
 }
 
 const sectionTitleStyles = css`
   @media screen and (min-width: 768px) {
     h3 {
-      font-size: 30px;
+      font-size: 22px;
+      line-height: 34px;
     }
 
     button {
       font-size: 20px;
     }
   }
+`
+
+const Container = styled.section``
+
+const Header = styled.div`
+  margin-bottom: 10px;
 `
 
 interface RightChildAsButtonProps {
@@ -50,7 +62,7 @@ function RightChildAsButton({ onClick, text, hide }: RightChildAsButtonProps) {
   if (hide) return
 
   return (
-    <Button
+    <RightChildButton
       type='link'
       onClick={onClick}
       highlight={false}
@@ -58,6 +70,12 @@ function RightChildAsButton({ onClick, text, hide }: RightChildAsButtonProps) {
     />
   )
 }
+
+const RightChildButton = styled(Button)`
+  @media screen and (min-width: 768px) {
+    font-size: 12px !important;
+  }
+`
 
 Section.RightChildAsButton = RightChildAsButton