+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.25.4
+
+- `Section.RightChildAsButton` font size was set as 14px for tablet screens
+- `Header` was added to `Section` with a margin-bottom
+- `Section` was enable to accept styles from styled
+
 # 1.25.3
 
 - Upgrade *defaultAfterStyles* prop for `Typography`